pyproject.toml
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml/#a-full-example
[project]
    authors         = [{ name = "Cédric HT" }]
    dependencies    = [
    "lightning>=2.6.1",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torchmetrics>=1.8.2",
    "torchvision>=0.25.0",
]
    description     = "Common classifier and classification data classes wrapped in Pytorch Lightning classes"
    license         = { file = "LICENSE" }
    name            = "lightning_wrappers"
    readme          = "README.md"
    requires-python = ">=3.11"
    version         = "0.0.0"
    [project.urls]
        "Bug Tracker" = "https://github.com/altaris/lightning_wrappers/issues"
        Changelog     = "https://github.com/altaris/lightning_wrappers/blob/main/CHANGELOG.md"
        Documentation = "https://github.com/altaris/lightning_wrappers"
        Homepage      = "https://github.com/altaris/lightning_wrappers"
        Repository    = "https://github.com/altaris/lightning_wrappers.git"
    [project.optional-dependencies]
        dev = ["pdoc>=14.7.0"]
# https://mypy.readthedocs.io/en/stable/config_file.html#config-file-format
[tool.mypy]
    check_untyped_defs     = true
    disallow_untyped_defs  = true
    ignore_missing_imports = true
    pretty                 = true
    python_version         = "3.11"
    strict_equality        = true
    warn_redundant_casts   = true
    warn_return_any        = true
    warn_unreachable       = true
    warn_unused_ignores    = true
[tool.pyright]
    include = ["lightning_wrappers"]
[tool.ruff]
    exclude = [
        ".bzr",
        ".direnv",
        ".eggs",
        ".git",
        ".git-rewrite",
        ".hg",
        ".ipynb_checkpoints",
        ".mypy_cache",
        ".nox",
        ".pants.d",
        ".pyenv",
        ".pytest_cache",
        ".pytype",
        ".ruff_cache",
        ".svn",
        ".tox",
        ".venv",
        ".vscode",
        "__pypackages__",
        "_build",
        "buck-out",
        "build",
        "dist",
        "node_modules",
        "site-packages",
        "template.py",
        "venv",
    ]
    line-length = 79
    target-version = "py311"
    [tool.ruff.format]
        docstring-code-format      = true
        docstring-code-line-length = "dynamic"
        indent-style               = "space"
        line-ending                = "auto"
        quote-style                = "double"
        skip-magic-trailing-comma  = false
    [tool.ruff.lint]
        # https://flake8.pycqa.org/en/latest/user/error-codes.html
        # https://pycodestyle.pycqa.org/en/latest/intro.html#error-codes
        # select = ["E4", "E7", "E9", "F"]
        fixable = ["ALL"]
        ignore = [
            "E722", # Do not use bare `except`
            "E731", # Do not assign a `lambda` expression, use a `def`
        ]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
timm = [
    "timm>=1.0.25",
]
transformers = [
    "transformers>=5.2.0",
]
